+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/analysis/AnalysesAjaxController.java
@@ -28,8 +28,7 @@ import ca.corefacility.bioinformatics.irida.ria.web.analysis.dto.AnalysesListReq
 import ca.corefacility.bioinformatics.irida.ria.web.analysis.dto.AnalysisModel;
 import ca.corefacility.bioinformatics.irida.ria.web.analysis.dto.AnalysisStateModel;
 import ca.corefacility.bioinformatics.irida.ria.web.analysis.dto.AnalysisTypeModel;
-import ca.corefacility.bioinformatics.irida.ria.web.models.TableModel;
-import ca.corefacility.bioinformatics.irida.ria.web.models.TableResponse;
+import ca.corefacility.bioinformatics.irida.ria.web.models.tables.TableResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.utilities.DateUtilities;
 import ca.corefacility.bioinformatics.irida.security.permissions.analysis.UpdateAnalysisSubmissionPermission;
 import ca.corefacility.bioinformatics.irida.service.AnalysisSubmissionService;
@@ -105,7 +104,7 @@ public class AnalysesAjaxController {
 	 * @throws IridaWorkflowNotFoundException thrown if the workflow cannot be found
 	 */
 	@RequestMapping("/list")
-	public TableResponse getPagedAnalyses(@RequestBody AnalysesListRequest analysesListRequest,
+	public TableResponse<AnalysisModel> getPagedAnalyses(@RequestBody AnalysesListRequest analysesListRequest,
 			@RequestParam(required = false, defaultValue = "false") Boolean admin,
 			@RequestParam(required = false) Long projectId, Locale locale) throws IridaWorkflowNotFoundException {
 
@@ -164,12 +163,12 @@ public class AnalysesAjaxController {
 		/*
 		UI cannot consume it as-is.  Format into something the UI will like use the the AnalysisModel
 		 */
-		List<TableModel> analyses = page.getContent()
+		List<AnalysisModel> analyses = page.getContent()
 				.stream()
 				.map(submission -> this.createAnalysisModel(submission, locale))
 				.collect(Collectors.toList());
 
-		return new TableResponse(analyses, page.getTotalElements());
+		return new TableResponse<AnalysisModel>(analyses, page.getTotalElements());
 	}
 
 	/**