+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/models/tables/TableResponse.java
@@ -1,20 +1,20 @@
-package ca.corefacility.bioinformatics.irida.ria.web.models;
+package ca.corefacility.bioinformatics.irida.ria.web.models.tables;
 
 import java.util.List;
 
 /**
  * Response sent when items are requested for a table.
  */
-public class TableResponse {
-	private List<TableModel> dataSource;
+public class TableResponse<T extends TableModel> {
+	private List<T> dataSource;
 	private Long total;
 
-	public TableResponse(List<TableModel> dataSource, Long total) {
+	public TableResponse(List<T> dataSource, Long total) {
 		this.dataSource = dataSource;
 		this.total = total;
 	}
 
-	public List<TableModel> getDataSource() {
+	public List<T> getDataSource() {
 		return dataSource;
 	}