+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes
 20.01 to 20.05
 --------------
 * [UI/Developer]: Removed webpack progress.
+* [UI]: Updated clients listing table to use `ant.design` tables.
 
 19.09 to 20.01
 --------------